n64: split recompiler in three files
