+++ ares/n64/CMakeLists.txt
@@ -157,6 +157,8 @@ ares_add_sources(
     cpu/interpreter.cpp
     cpu/memory.cpp
     cpu/recompiler.cpp
+    cpu/recompiler-fpu.cpp
+    cpu/recompiler-ipu.cpp
     cpu/serialization.cpp
     cpu/tlb.cpp
 )
